Shuffle around types, add traceItemMeta to based render function baggage
