+++ static/app/views/explore/logs/utils.tsx
@@ -3,12 +3,10 @@ import * as Sentry from '@sentry/react';
 import * as qs from 'query-string';
 
 import type {ApiResult} from 'sentry/api';
-import {getTooltipText as getAnnotatedTooltipText} from 'sentry/components/events/meta/annotatedText/utils';
 import {t} from 'sentry/locale';
 import type {PageFilters} from 'sentry/types/core';
 import type {TagCollection} from 'sentry/types/group';
 import type {Organization} from 'sentry/types/organization';
-import type {Project} from 'sentry/types/project';
 import {defined} from 'sentry/utils';
 import type {EventsMetaType} from 'sentry/utils/discover/eventView';
 import {
@@ -34,8 +32,6 @@ import {LOGS_SORT_BYS_KEY} from 'sentry/views/explore/contexts/logs/sortBys';
 import {Mode} from 'sentry/views/explore/contexts/pageParamsContext/mode';
 import {SavedQuery} from 'sentry/views/explore/hooks/useGetSavedQueries';
 import type {
-  TraceItemAttributeMeta,
-  TraceItemDetailsMeta,
   TraceItemDetailsResponse,
   TraceItemResponseAttribute,
 } from 'sentry/views/explore/hooks/useTraceItemDetails';
@@ -507,77 +503,4 @@ export function ourlogToJson(ourlog: TraceItemDetailsResponse | undefined): stri
     }
   }
   return JSON.stringify(copy, null, 2);
-}
-
-/**
- * Metadata about trace item attributes.
- *
- * This can be used to extract additional information about attributes
- * like remarks (e.g. why a value was redacted).
- */
-export class TraceItemMetaInfo {
-  private static readonly META_PATH_CURRENT = '';
-
-  private meta: TraceItemDetailsMeta;
-
-  constructor(meta: TraceItemDetailsMeta) {
-    this.meta = meta;
-  }
-
-  getAttributeMeta(attribute: string): TraceItemAttributeMeta | undefined {
-    return this.meta?.[attribute]?.meta?.value?.[TraceItemMetaInfo.META_PATH_CURRENT];
-  }
-
-  getRemarks(attribute: string): RemarkObject[] {
-    const attributeMeta = this.getAttributeMeta(attribute);
-    if (!attributeMeta?.rem?.length) {
-      return [];
-    }
-
-    const properlyFormedRemarks = attributeMeta.rem.filter(r => r.length >= 2); // Defensive against unexpected length remarks.
-
-    return properlyFormedRemarks.map((rem): RemarkObject => {
-      const [ruleId, type, rangeStart, rangeEnd] = rem;
-      return {
-        ruleId: String(ruleId),
-        type: String(type),
-        rangeStart: Number(rangeStart),
-        rangeEnd: Number(rangeEnd),
-      };
-    });
-  }
-
-  hasRemarks(attribute: string): boolean {
-    const attributeMeta = this.getAttributeMeta(attribute);
-    return (attributeMeta?.rem?.length ?? 0) > 0;
-  }
-
-  static getTooltipText(
-    attribute: string,
-    meta: TraceItemDetailsMeta,
-    organization?: Organization,
-    project?: Project
-  ): string | React.ReactNode | null {
-    const metaInfo = new TraceItemMetaInfo(meta);
-    const remarks = metaInfo.getRemarks(attribute);
-
-    const firstRemark = remarks[0];
-    if (!firstRemark) {
-      return null;
-    }
-
-    return getAnnotatedTooltipText({
-      remark: firstRemark.type,
-      rule_id: firstRemark.ruleId,
-      organization,
-      project,
-    });
-  }
-}
-
-interface RemarkObject {
-  rangeEnd: number;
-  rangeStart: number;
-  ruleId: string;
-  type: string;
 }
